+++ package.json
@@ -9,14 +9,8 @@
     "react-dom": "^16.2.0",
     "react-scripts": "1.0.17"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  },
-  "description": "This project was bootstrapped with [Create React App](https://github.com/facebookincubator/create-react-app).",
-  "main": "web_server.js",
+  "description": "This project sets up a Node server.",
+  "main": "server.js",
   "devDependencies": {},
   "repository": {
     "type": "git",
