--- src/App.js
@@ -1,16 +1,0 @@
-import React, { Component } from 'react';
-import './App.css';
-
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <header className="App-header">
-          <h1 className="App-title">Welcome to Message-app</h1>
-        </header>
-      </div>
-    );
-  }
-}
-
-export default App;
